Edición de listas, media queries, diseño mejorado

+++ src/app/main-page/main-page.component.ts
@@ -2,15 +2,16 @@ import { Component, computed, inject } from '@angular/core';
 import { UserInteractionsService } from '../services/user-interactions.service';
 import { OmdbApiService } from '../services/omdb-api.service';
 import { Watched } from '../models/watched.model';
-import { DatePipe } from '@angular/common';
+import { DatePipe, NgClass } from '@angular/common';
 import { DarkModeService } from '../services/dark-mode.service';
 import { ListsService } from '../services/lists-modal.service';
 import { List } from '../models/list.model';
 import { ContentDataModel } from '../models/content-data.model';
+import { ConfirmModalService } from '../services/confirm-modal.service';
 
 @Component({
   selector: 'app-main-page',
-  imports: [DatePipe],
+  imports: [DatePipe, NgClass],
   templateUrl: './main-page.component.html',
   styleUrl: './main-page.component.css'
 })
@@ -19,11 +20,14 @@ export class MainPageComponent {
   userInteractionService = inject(UserInteractionsService);
   darkModeService = inject(DarkModeService);
   listsService = inject(ListsService);
+  confirmService = inject(ConfirmModalService);
   favorites: ContentDataModel[] = [];
   watched: Watched[] = [];
   lists: List[] = [];
 
   buttonSvgFill = computed(() => (this.darkModeService.darkModeOn() ? '#6ea8fe' : '#052c65'));
+  separatorHrFill = computed(() => (this.darkModeService.darkModeOn() ? 'topSeparatorDark' : 'topSeparatorLight'));
+  badgeBackground = computed(() => (this.darkModeService.darkModeOn() ? 'text-bg-dark' : 'text-bg-info'));
 
   ngOnInit() {
     this.favorites = this.userInteractionService.getFavorites();
@@ -46,12 +50,20 @@ export class MainPageComponent {
     this.watched = this.userInteractionService.sortWatched();
   }
 
-  removeEntry(imdbID: string, date: string) {
-    this.userInteractionService.removeWatched(imdbID, date);
+  removeEntry(imdbID: string, title: string, date: string) {
+    this.confirmService.confirm('¿Deseas eliminar "' + title +  '" de Actividad?', (confirmed) => {
+      if (confirmed) {
+        this.userInteractionService.removeWatched(imdbID, date);
+      }
+    });
   }
 
-  removeFavorite(imdbID: string) {
-    this.userInteractionService.removeFavorite(imdbID);
+  removeFavorite(imdbID: string, title: string) {
+    this.confirmService.confirm('¿Deseas eliminar "' + title + '" de Favoritos?', (confirmed) => {
+      if (confirmed) {
+        this.userInteractionService.removeFavorite(imdbID);
+      }
+    });
   }
 
   showAddListForm() {
@@ -67,7 +79,15 @@ export class MainPageComponent {
     this.listsService.contentToAdd(content);
   }
 
-  removeList(listID: string) {
-    this.userInteractionService.removeList(listID);
+  removeList(listID: string, name: string) {
+    this.confirmService.confirm('¿Deseas eliminar la lista "' + name + '"?', (confirmed) => {
+      if (confirmed) {
+        this.userInteractionService.removeList(listID);
+      }
+    });
+  }
+
+  editList(list: List) {
+    this.listsService.editList(list);
   }
 }
